Criar pedido
Implementação da funcionalidade de criação de pedido.

+++ src/main/java/com/perisatto/fiapprj/menuguru/order/domain/model/OrderStatus.java
@@ -1,22 +1,22 @@
 package com.perisatto.fiapprj.menuguru.order.domain.model;
 
 public enum OrderStatus {
-	RECEBIDO(1),
-	EM_PREPARACAO(2),
-	PRONTO(3),
-	FINALIZADO(4);
+	RECEBIDO(1L),
+	EM_PREPARACAO(2L),
+	PRONTO(3L),
+	FINALIZADO(4L);
 	
-	private Integer id;
+	private Long id;
 	
-	OrderStatus(Integer id) {
+	OrderStatus(Long id) {
 		this.id = id;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 	
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 }
